[DXGI] Change DXGIError to class using SharpDX.Result fields

+++ Source/SharpDX.DXGI/Adapter.cs
@@ -103,7 +103,7 @@ namespace SharpDX.DXGI
                 }
                 catch (SharpDXException exception)
                 {
-                    if (exception.ResultCode.Code == (int)DXGIError.NotFound)
+                    if (exception.ResultCode.Code == DXGIError.NotFound)
                         break;
                     throw;
                 }
