+++ Source/SharpDX.DXGI/Factory1.cs
@@ -49,7 +49,7 @@ namespace SharpDX.DXGI
                 }
                 catch (SharpDXException exception)
                 {
-                    if (exception.ResultCode.Code == (int) DXGIError.NotFound)
+                    if (exception.ResultCode.Code == DXGIError.NotFound)
                         break;
                     throw;
                 }
